fix: altera componente da header
