+++ frontend/src/pages/search/index.js
@@ -3,8 +3,8 @@ import "./styles.css";
 
 import axios from "axios";
 
+// import Header from "../../components/header";
 import Header from "../../components/header";
-import HeaderDesktop from "../../components/headerDesktop";
 import Pagination from "../../components/pagination";
 
 import { useLocation, Link } from "react-router-dom";
@@ -145,12 +145,9 @@ const Search = () => {
 
   return (
     <>
-      {(window.innerWidth > 760) ?
-        <HeaderDesktop />
-        :
+
 
         <Header />
-      }
 
       <div className="search-container">
         <div className="search-content">
